+++ tarantool/src/ffi/helper.rs
@@ -82,6 +82,105 @@ macro_rules! static_assert {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
+// offset_of!
+////////////////////////////////////////////////////////////////////////////////
+
+/// Returns an offset of the struct or tuple member in bytes.
+///
+/// Returns a constant which can be used at compile time.
+///
+/// # Example
+/// ```rust
+/// # use tarantool::offset_of;
+/// #[repr(C)]
+/// struct MyStruct { a: u8, b: u8 }
+/// assert_eq!(offset_of!(MyStruct, a), 0);
+/// assert_eq!(offset_of!(MyStruct, b), 1);
+///
+/// // Also works with tuples:
+/// assert_eq!(offset_of!((i32, i32), 0), 0);
+/// assert_eq!(offset_of!((i32, i32), 1), 4);
+/// ```
+#[macro_export]
+macro_rules! offset_of {
+    ($type:ty, $field:tt) => {{
+        const RESULT: usize = unsafe {
+            let dummy = ::core::mem::MaybeUninit::<$type>::uninit();
+            let dummy_ptr = dummy.as_ptr();
+            let field_ptr = ::std::ptr::addr_of!((*dummy_ptr).$field);
+
+            let field_ptr = field_ptr.cast::<u8>();
+            let dummy_ptr = dummy_ptr.cast::<u8>();
+            field_ptr.offset_from(dummy_ptr) as usize
+        };
+        RESULT
+    }};
+}
+
+/// Returns size of type, or type's field.
+///
+/// Returns a constant which can be used at compile time.
+///
+/// # Example
+/// ```rust
+/// # use tarantool::size_of;
+/// #[repr(C)]
+/// struct MyStruct { a: u8, b: u16 }
+/// assert_eq!(size_of!(MyStruct, a), 1);
+/// assert_eq!(size_of!(MyStruct, b), 2);
+///
+/// // Also works with tuples:
+/// assert_eq!(size_of!((i32, i32), 0), 4);
+/// assert_eq!(size_of!((i32, i32), 1), 4);
+/// ```
+#[macro_export]
+macro_rules! size_of {
+    ($type:ty) => {
+        ::std::mem::size_of::<$type>()
+    };
+    ($type:ty, $field:tt) => {{
+        const RESULT: usize = unsafe {
+            let dummy = ::core::mem::MaybeUninit::<$type>::uninit();
+            let dummy_ptr = dummy.as_ptr();
+            let field_ptr = ::std::ptr::addr_of!((*dummy_ptr).$field);
+
+            const fn size_of_val<T>(_: *const T) -> usize {
+                ::std::mem::size_of::<T>()
+            }
+
+            size_of_val(field_ptr)
+        };
+        RESULT
+    }};
+}
+
+const _TEST_OFFSET_AND_SIZE_OF: () = {
+    #[repr(C)]
+    struct MyStruct {
+        a: u8,
+        b: u16,
+        c: u32,
+        d: u64,
+    }
+
+    assert!(offset_of!(MyStruct, a) == 0);
+    assert!(offset_of!(MyStruct, b) == 2);
+    assert!(offset_of!(MyStruct, c) == 4);
+    assert!(offset_of!(MyStruct, d) == 8);
+
+    assert!(size_of!(MyStruct, a) == 1);
+    assert!(size_of!(MyStruct, b) == 2);
+    assert!(size_of!(MyStruct, c) == 4);
+    assert!(size_of!(MyStruct, d) == 8);
+
+    assert!(offset_of!((i32, i32), 0) == 0);
+    assert!(offset_of!((i32, i32), 1) == 4);
+
+    assert!(size_of!((i32, i32), 0) == 4);
+    assert!(size_of!((i32, i32), 1) == 4);
+};
+
+////////////////////////////////////////////////////////////////////////////////
 // define_dlsym_reloc!
 ////////////////////////////////////////////////////////////////////////////////